Create Non-repeating Character
Given a string S consisting of lowercase Latin Letters. Return the first non-repeating character in S. If there is no non-repeating character, return '$'.


Input:
S = hello
Output: h
Explanation: In the given string, the first character which is non- repeating is h, as it appears first and there is no other 'h' in the string.